refactor task logic
